+++ src/utils/wslUtils.ts
@@ -11,12 +11,20 @@ export function useWsl(): boolean {
 }
 
 export async function toWslPath(path: string): Promise<string> {
-    return (await executeCommand("wsl", ["wslpath", "-u", `"${path.replace(/\\/g, "/")}"`])).trim();
+    return (await executeCommand(
+        "wsl",
+        ["wslpath", "-u", path.replace(/\\/g, "/")],
+        { shell: false },
+    )).trim();
 }
 
 export async function toWinPath(path: string): Promise<string> {
     if (path.startsWith("\\mnt\\")) {
-        return (await executeCommand("wsl", ["wslpath", "-w", `"${path.replace(/\\/g, "/").substr(0, 6)}"`])).trim() + path.substr(7);
+        return (await executeCommand(
+            "wsl",
+            ["wslpath", "-w", path.replace(/\\/g, "/").substr(0, 6)],
+            { shell: false },
+        )).trim() + path.substr(7);
     }
-    return (await executeCommand("wsl", ["wslpath", "-w", "/"])).trim() + path;
+    return (await executeCommand("wsl", ["wslpath", "-w", "/"], { shell: false })).trim() + path;
 }